renamed strings.py to prompts.py as it only contains prompts
