+++ src/ai_hawk/llm/prompts.py
